requirements.txt
llama-index==0.8.52
openai==0.28.1
pyautogen==0.1.14
docker==6.1.3
pyautogen[retrievechat]==0.1.14
llama-hub==0.0.43
wikipedia==1.4.0
streamlit